Add files via upload
Criminal cases in Bangladesh are rising at an alarming rate. I wanted to create a simple dashboard for the regular public to see the severity of the issue. The dashboard allows you to choose crimes by category, year and police units. The raw data was taken from kaggle.com, cleaned in SQL and visualized in Power BI